+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### New Features
 ### Bug Fixes
 * Remove explicit version on `micrometer-registry-prometheus` dependency. (#563)
+* Send a `TaskNotificationType` to worker through a 2XX HTTP status code. (#564)
 ### Quality
 ### Dependency Upgrades